update: fine tune colors & update permissions on click

+++ src/features/nodes/hooks/tree/node/actions/useNodeDelete.tsx
@@ -3,6 +3,7 @@ import { reloadBranch } from '../../../../../branch/branches.thunks';
 import useBranchContext from '../../../../../branch/hooks/useBranchContext';
 import { deleteFromState } from '../../../../nodes.actions';
 import { deleteNode } from '../../../../nodes.thunks';
+import useAuthorizeNodeAction from '../useAuthorizeNodeAction';
 import useNodeContext from '../useNodeContext';
 import { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
@@ -15,8 +16,13 @@ export default function useNodeDelete() {
         treeRootId, branchId, id, rootId, isTmp,
     } = useNodeContext();
     const navigate = useNavigate();
+    const authorizeNodeAction = useAuthorizeNodeAction();
 
     return useCallback(async () => {
+        if (!authorizeNodeAction()) {
+            return;
+        }
+
         if (isTmp) {
             dispatch(deleteFromState({
                 branchId,
@@ -34,5 +40,5 @@ export default function useNodeDelete() {
             }
         }
         if (treeRootId === id) navigate('/nodes');
-    }, [branchId, dispatch, id, rootId, isBranch, isTmp, navigate, treeRootId]);
+    }, [authorizeNodeAction, isTmp, treeRootId, id, navigate, dispatch, branchId, rootId, isBranch]);
 }
